Minor improvement of code, moved number generators to "utils"
