+++ datageneration/tests/test_retrieve_combination.py
@@ -10,7 +10,7 @@ Execute it as follows: python -m unittest datageneration.tests.test_retrieve_com
 
 class TestCombinationRetriever(unittest.TestCase):
     def setUp(self):
-        self.retriever = CombinationRetriever(source='datageneration/tests/data/Primary_Keys_filtered10.xlsx',
+        self.retriever = CombinationRetriever(source='datageneration/tests/data/Primary_Keys_test10.xlsx',
                                               prop_limit=100)
 
     def compare_tags(self, tested_tags, all_tags):
